infra/rag/requirements.txt
PyMuPDF
openai
tenacity
langchain
tiktoken
azure-search-documents==11.6.0b4
azure-core
azure-identity
loguru
python-docx 
markdownify
#azure content understanding client is not available in pypi - using preview code in python folder
